requirements.txt
# API
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
httpx==0.27.2

# LLM & Agent
langchain==0.3.0
langchain-community==0.3.0
langchain-ollama==0.2.0
langchain-text-splitters==0.3.0
langgraph==0.2.28

# Vector DB
qdrant-client==1.11.0

# Cache
redis==5.1.0

# Database
sqlalchemy==2.0.35
psycopg2-binary==2.9.9

# Document processing
unstructured[all-docs]==0.15.13

# Web search
tavily-python==0.5.0

# Utilities
python-dotenv==1.0.1
pydantic==2.9.2
